CORE-1184: Fix JSON deserialization, rename as Blockset brd impls
