+++ WalletKitJava/WalletKit/src/main/java/com/blockset/walletkit/systemclient/brd/BlocksetSubscription.java
@@ -7,26 +7,28 @@
  */
 package com.blockset.walletkit.systemclient.brd;
 
-import com.blockset.walletkit.SystemClient;
 import com.blockset.walletkit.SystemClient.SubscriptionEndpoint;
+import com.blockset.walletkit.SystemClient.Subscription;
 import com.blockset.walletkit.SystemClient.SubscriptionCurrency;
 import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-public class Subscription implements SystemClient.Subscription {
+public class BlocksetSubscription implements Subscription {
 
     // creators
 
     @JsonCreator
-    public static Subscription create(@JsonProperty("subscription_id") String id,
-                                      @JsonProperty("device_id") String deviceId,
-                                      @JsonProperty("endpoint") SubscriptionEndpoint endpoint,
-                                      @JsonProperty("currencies") List<SubscriptionCurrency> currencies) {
-        return new Subscription(
+    public static BlocksetSubscription create(@JsonProperty("subscription_id") String id,
+                                              @JsonProperty("device_id") String deviceId,
+                                              @JsonProperty("endpoint") BlocksetSubscriptionEndpoint endpoint,
+                                              @JsonProperty("currencies") List<BlocksetSubscriptionCurrency> currencies) {
+        return new BlocksetSubscription(
                 checkNotNull(id),
                 checkNotNull(deviceId),
                 checkNotNull(endpoint),
@@ -38,13 +40,13 @@ public class Subscription implements SystemClient.Subscription {
 
     private final String id;
     private final String deviceId;
-    private final SubscriptionEndpoint endpoint;
-    private final List<SubscriptionCurrency> currencies;
+    private final BlocksetSubscriptionEndpoint endpoint;
+    private final List<BlocksetSubscriptionCurrency> currencies;
 
-    private Subscription(String subscriptionId,
-                         String deviceId,
-                         SubscriptionEndpoint endpoint,
-                         List<SubscriptionCurrency> currencies) {
+    private BlocksetSubscription(String subscriptionId,
+                                 String deviceId,
+                                 BlocksetSubscriptionEndpoint endpoint,
+                                 List<BlocksetSubscriptionCurrency> currencies) {
         this.id = subscriptionId;
         this.deviceId = deviceId;
         this.endpoint = endpoint;
@@ -53,23 +55,33 @@ public class Subscription implements SystemClient.Subscription {
 
     // getters
 
+    @Override
     @JsonProperty("subscription_id")
     public String getId() {
         return id;
     }
 
+    @Override
     @JsonProperty("device_id")
     public String getDevice() {
         return deviceId;
     }
 
+    @Override
     @JsonProperty("endpoint")
     public SubscriptionEndpoint getEndpoint() {
         return endpoint;
     }
 
-    @JsonProperty("currencies")
+    @Override
+    @JsonIgnore
     public List<SubscriptionCurrency> getCurrencies() {
+        return new ArrayList<SubscriptionCurrency>(getBlocksetCurrencies());
+    }
+
+    @JsonProperty("currencies")
+    private List<BlocksetSubscriptionCurrency> getBlocksetCurrencies() {
         return currencies;
     }
+
 }
