+++ WalletKitJava/WalletKit/src/main/java/com/blockset/walletkit/systemclient/brd/BlocksetSubscriptionEvent.java
@@ -6,7 +6,8 @@
  * See the CONTRIBUTORS file at the project root for a list of contributors.
  */
 package com.blockset.walletkit.systemclient.brd;
-import com.blockset.walletkit.SystemClient;
+
+import com.blockset.walletkit.SystemClient.SubscriptionEvent;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.primitives.UnsignedInteger;
@@ -14,14 +15,14 @@ import java.util.List;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-public class SubscriptionEvent implements SystemClient.SubscriptionEvent {
+public class BlocksetSubscriptionEvent implements SubscriptionEvent {
 
     // creators
 
     @JsonCreator
-    public static SubscriptionEvent create(@JsonProperty("name") String name,
-                                           @JsonProperty("confirmations") List<UnsignedInteger> confirmations) {
-        return new SubscriptionEvent(
+    public static BlocksetSubscriptionEvent create(@JsonProperty("name") String name,
+                                                   @JsonProperty("confirmations") List<UnsignedInteger> confirmations) {
+        return new BlocksetSubscriptionEvent(
                 checkNotNull(name),
                 checkNotNull(confirmations)
         );
@@ -32,19 +33,21 @@ public class SubscriptionEvent implements SystemClient.SubscriptionEvent {
     private final String name;
     private final List<UnsignedInteger> confirmations;
 
-    private SubscriptionEvent(String name,
-                              List<UnsignedInteger> confirmations) {
+    private BlocksetSubscriptionEvent(String name,
+                                      List<UnsignedInteger> confirmations) {
         this.name = name;
         this.confirmations = confirmations;
     }
 
     // getters
 
+    @Override
     @JsonProperty("name")
     public String getName() {
         return name;
     }
 
+    @Override
     @JsonProperty("confirmations")
     public List<UnsignedInteger> getConfirmations() {
         return confirmations;
