+++ WalletKitJava/WalletKit/src/main/java/com/blockset/walletkit/systemclient/brd/BlocksetTransactionIdentifier.java
@@ -6,24 +6,25 @@
  * See the CONTRIBUTORS file at the project root for a list of contributors.
  */
 package com.blockset.walletkit.systemclient.brd;
+
 import android.support.annotation.Nullable;
 
-import com.blockset.walletkit.SystemClient;
+import com.blockset.walletkit.SystemClient.TransactionIdentifier;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Optional;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-public class TransactionIdentifier implements SystemClient.TransactionIdentifier {
+public class BlocksetTransactionIdentifier implements TransactionIdentifier {
     // creator
 
     @JsonCreator
-    public static TransactionIdentifier create(@JsonProperty("transaction_id") String id,
-                                               @JsonProperty("identifier") String identifier,
-                                               @JsonProperty("hash") @Nullable String hash,
-                                               @JsonProperty("blockchain_id") String blockchainId) {
-        return new TransactionIdentifier(
+    public static BlocksetTransactionIdentifier create(@JsonProperty("transaction_id") String id,
+                                                       @JsonProperty("identifier") String identifier,
+                                                       @JsonProperty("hash") @Nullable String hash,
+                                                       @JsonProperty("blockchain_id") String blockchainId) {
+        return new BlocksetTransactionIdentifier(
                 checkNotNull(id),
                 checkNotNull(identifier),
                 hash,
@@ -38,10 +39,10 @@ public class TransactionIdentifier implements SystemClient.TransactionIdentifier
     private final @Nullable String hash;
     private final String blockchainId;
 
-    private TransactionIdentifier(String id,
-                                  String identifier,
-                                  @Nullable String hash,
-                                  String blockchainId) {
+    private BlocksetTransactionIdentifier(String id,
+                                          String identifier,
+                                          @Nullable String hash,
+                                          String blockchainId) {
         this.id = id;
         this.identifier = identifier;
         this.hash = hash;
@@ -49,23 +50,28 @@ public class TransactionIdentifier implements SystemClient.TransactionIdentifier
     }
     // getters
 
+    @Override
     @JsonProperty("transaction_id")
     public String getId() {
         return id;
     }
 
-    @JsonProperty("identifier")
-    public String getIdentifier() {
-        return identifier;
+    @Override
+    @JsonProperty("blockchain_id")
+    public String getBlockchainId() {
+        return blockchainId;
     }
 
+    @Override
     @JsonProperty("hash")
     public Optional<String> getHash() {
         return Optional.fromNullable (hash);
     }
 
-    @JsonProperty("blockchain_id")
-    public String getBlockchainId() {
-        return blockchainId;
+    @Override
+    @JsonProperty("identifier")
+    public String getIdentifier() {
+        return identifier;
     }
+
 }
