+++ WalletKitJava/WalletKitBRD/src/main/java/com/blockset/walletkit/brd/Blockchains.java
@@ -12,6 +12,7 @@ import java.util.List;
 import java.util.Locale;
 
 import com.blockset.walletkit.SystemClient.CurrencyDenomination;
+import com.blockset.walletkit.systemclient.brd.BlocksetCurrencyDenomination;
 import com.google.common.collect.ImmutableList;
 import com.google.common.primitives.UnsignedInteger;
 
@@ -25,25 +26,18 @@ final class Blockchains {
     static final String ADDRESS_BRD_TESTNET = "0x7108ca7c4718efa810457f228305c9c71390931a";
 
     /* package */
-    static List<CurrencyDenomination> makeCurrencyDemominationsErc20 (String code, UnsignedInteger decimals) {
+    static List<BlocksetCurrencyDenomination> makeCurrencyDemominationsErc20 (String code, UnsignedInteger decimals) {
         String name = code.toUpperCase(Locale.ROOT);
         code = code.toLowerCase(Locale.ROOT);
 
-        // TODO: confirm that its OK to poke through the inscrutability of
-        //       the SystemClient interface and reach creation method for BRD/blockset
-        //       in order to get something concrete. This should be OK? since this java
-        //       is a BRD package and so might be assumed to be accessible to the
-        //       JSON/blockset derivations. Otherwise the SystemClient interface
-        //       will also have to impose 'create' methods that each implemenation
-        //       provides
         return ImmutableList.of(
-                com.blockset.walletkit.systemclient.brd.CurrencyDenomination.create(
+                BlocksetCurrencyDenomination.create(
                         String.format(Locale.ROOT, "%s Token INT", name),
                         String.format(Locale.ROOT, "%si", code),
                         UnsignedInteger.ZERO,
                         String.format(Locale.ROOT, "%si", code)
                 ),
-                com.blockset.walletkit.systemclient.brd.CurrencyDenomination.create(
+                BlocksetCurrencyDenomination.create(
                         String.format(Locale.ROOT, "%s Token", name),
                         code,
                         decimals,
@@ -51,6 +45,4 @@ final class Blockchains {
                 )
         );
     }
-
-
 }
